% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird.R
\name{ebird}
\alias{ebird}
\title{Return a remote data.frame connection to your local ebird database}
\usage{
ebird(conn = ebird_conn())
}
\arguments{
\item{conn}{a connection to the local ebird database, see \verb{[ebird_conn]}.}
}
\description{
Return a remote data.frame connection to your local ebird database
}
\examples{

Sys.setenv("BIRDDB_HOME"=tempdir())
tar <- ebird_sample_data()
import_ebird(tar)
df <- ebird()

}
