% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_remote.R
\name{ebird_remote}
\alias{ebird_remote}
\title{ebird remote}
\usage{
ebird_remote(
  version = "Jul-2021",
  bucket = "ebird",
  to_duckdb = FALSE,
  host = "minio.thelio.carlboettiger.info",
  ...
)
}
\arguments{
\item{version}{eBird snapshot date}

\item{bucket}{eBird bucket name (including region)}

\item{to_duckdb}{Return a remote duckdb connection or arrow connection?
Note that leaving as FALSE may be faster but is limited to the dplyr-style
operations supported by \link{arrow} alone.}

\item{...}{additional parameters passed to the s3_bucket() (e.g. for remote
access to independently hosted buckets)}
}
\description{
Connect to an eBird snapshot remote. Can be much faster than downloading
for one-off use or when using the package from a server in the same region
as the data.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
