% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_data_dir.R
\name{ebird_sample_data}
\alias{ebird_sample_data}
\title{Provide path to a small subset of ebird data}
\usage{
ebird_sample_data()
}
\description{
Sample data is based on the official sample download from July 2021.
Sample is modified slightly to better match the official full download:
archive uses \code{tar} format instead of \code{zip}, and the \verb{ebd_*} table is
compressed with \code{gzip}, as in the full tar archive distributions.
This data should be used for testing purposes only.
}
\examples{
ebird_sample_data()
}
