% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\name{sample_data}
\alias{sample_data}
\alias{sample_checklist_data}
\alias{sample_observation_data}
\title{Provide path to a small subset of eBird data}
\usage{
sample_checklist_data()

sample_observation_data()
}
\value{
The path to the sample tar archive file.
}
\description{
These small sample dataset consists of all observations from Hong Kong in the
year 2012. Sample files are provided for checklist and observation data, both
packaged as tar archive files to mimic the format of the eBird Basic Dataset
download.
}
\examples{
sample_checklist_data()
sample_observation_data()
}
