% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_conn.R
\name{ebird_conn}
\alias{ebird_conn}
\title{A \verb{[DBI]}-style database connection to the imported eBird data}
\usage{
ebird_conn()
}
\description{
Returns a database connection to the local eBird dataset.
}
\examples{

conn <- ebird_conn()
DBI::dbListTables(conn)

}
