% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebird_data_dir.R
\name{ebird_data_dir}
\alias{ebird_data_dir}
\title{ebird data dir}
\usage{
ebird_data_dir()
}
\description{
Show the location ebird uses to store data. The default location is that
chosen by R based on your OS, see \verb{[tools::R_user_dir]}.
Alternately, users can configure a different permanent storage location by
setting their desired path in the environmental variable BIRDDB_HOME.
This may be desirable when multiple users of the same machine or server want
to access a single copy of the ebird data.
}
