% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ebird.R
\name{import_ebird}
\alias{import_ebird}
\title{import ebird}
\usage{
import_ebird(tarfile)
}
\arguments{
\item{tarfile}{path to the copy of \verb{ebd_rel<DATE>.tar} file you downloaded}
}
\description{
Given a path to the ebird tarfile, this function will extract and import the
tar archive into a parquet-based database in your \code{\link[=ebird_data_dir]{ebird_data_dir()}}
}
